pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jobmatch-local"
version = "0.1.0"
description = "Local personal-use job matching tool built with NiceGUI"
requires-python = ">=3.12"
dependencies = [
  "beautifulsoup4>=4.14.3",
  "httpx>=0.28.1",
  "nicegui>=3.10.0",
  "playwright>=1.59.0",
  "pymupdf>=1.27.2.3",
  "python-dateutil>=2.9.0.post0",
  "python-docx>=1.2.0",
  "scikit-learn>=1.8.0",
  "sentence-transformers>=5.4.1",
  "SQLAlchemy>=2.0.49",
]

[project.optional-dependencies]
dev = ["pytest>=9.0.3"]

[project.scripts]
jobmatch = "app.cli:main"
jobmatch-ui = "app.ui.main:run"

[tool.setuptools.packages.find]
include = ["app*"]

